Converted unit price to float

+++ lib/loader.rb
@@ -63,7 +63,7 @@ class Loader
     file.each do |line|
       database.execute("insert into items values (?, ?, ?, ?, ?, ?, ?, ?, ?)",
                        line[:id].to_i, line[:name], line[:description], 
-                       line[:unit_price].to_i/100, line[:merchant_id].to_i,
+                       line[:unit_price].to_f/100, line[:merchant_id].to_i,
                        line[:created_at], line[:updated_at],
                        line[:created_at][0, 19], line[:updated_at][0, 19])
     end
