+++ lib/sales_engine/item_record.rb
@@ -6,7 +6,7 @@ module SalesEngine
          id = row[0].to_i
          name = row[1]
          description = row[2]
-         unit_price = row[3].to_i
+         unit_price = row[3].to_f
          merchant_id = row[4].to_i
          created_at = row[5]
          updated_at = row[6]
